Change initializing hierarchy long to int, and initialize the array

+++ src/PyBind11/PyBind11Demo.h
@@ -9,7 +9,7 @@ int PyBind11CallTestScript();
 int PyBind11SetParameters(struct yt_param_yt* yt_param_ptr);
 int PyBind11SetUserParameterInt(const char* key, int value);
 int PyBind11SetFields(struct yt_field* yt_field_ptr, int len);
-int PyBind11InitHier(long num_grids);
+int PyBind11InitHier(int num_grids);
 int PyBind11Run(const char* inline_script, const char* function);
 #endif  // #ifdef USE_PYBIND11
